lf land: stage uncommitted changes
