+++ .github/workflows/release.yml
@@ -80,8 +80,7 @@ jobs:
       - name: Install create-dmg
         run: brew install create-dmg
       - name: Build, sign, and notarize DMG
-        run: python3 scripts/concerto-dev.py release
-        working-directory: ${{ github.workspace }}
+        run: python3 scripts/release-concerto.py
       - name: Upload DMG to R2
         run: |
           uv run --with boto3 python3 - <<'SCRIPT'
